scripts/requirements.txt
python-whois
dnspython
requests
beautifulsoup4
colorama
ipinfo
